Dimension names

Extract names of various dimensions.

dim_names(x)

dim_names(x) <- value

set_dim_names(x, n, nms)

set_row_names(x, nms)

set_col_names(x, nms)

row_names(x)

col_names(x)

n_dim_names(x, n)

Arguments

x
The object to extract the dimension names for.

value
The new dimension names to use for x. This is a list of character vectors.

n
The n-th dimension to use.

nms
A character vector of new dimension names for the n-th dimension.

Details

Unlike dimnames() which can return NULL and contain elements that are NULL, dim_names() always returns a list the same length as the dimensionality of x. If any dimensions do not have names, character(0) is returned for that element of the list. This results in a type stable result: a list where the elements are character vectors.

A vector is treated as a 1 column matrix (so, 2 dimensions) and dim_names() will return the names of the vector as the row names, if it has any.

Examples

x <- mtrx(x = 1:5, y = 6:10, row_names = letters[1:5])
dim_names(x)

#> [[1]]
#> [1] "a" "b" "c" "d" "e"
#> 
#> [[2]]
#> [1] "x" "y"
#> 

# 2D object, so 2 sets of dim names
dim_names(rray())

#> [[1]]
#> character(0)
#> 
#> [[2]]
#> character(0)
#> 

# 3D object, so 3 sets of dim names
dim_names(rray(1, dim = c(1,1,1)))

#> [[1]]
#> character(0)
#> 
#> [[2]]
#> character(0)
#> 
#> [[3]]
#> character(0)
#> 

# Vectors are 1 column matrices (2D)
dim_names(1:5)

#> [[1]]
#> character(0)
#> 
#> [[2]]
#> character(0)
#> 

vec <- c(x = 1, y = 2)
dim_names(vec)

#> [[1]]
#> [1] "x" "y"
#> 
#> [[2]]
#> character(0)
#> 

# You can add dim names using set_dim_names()
# and the pipe operator
library(magrittr)
rray(1, c(1, 2, 1)) %>%
  set_dim_names(1, "r1") %>%
  set_dim_names(2, c("c1", "c2")) %>%
  set_dim_names(3, "3rd dim")

#> <vctrs_rray<double>[,2,1][2]>
#> , , 3rd dim
#> 
#>    c1 c2
#> r1 1  1 
#>