Build a mtrx object

Helpful constructor for building mtrx objects. Pass in named vectors, where each vector represents a single column in the mtrx. Length 1 vectors are recycled.

mtrx(..., row_names = character())

Arguments

...
Vector that represent the columns of the mtrx. These must be the same length, or can be length 1 in which case the vector will be recycled. Vectors are combined with vec_c() which performs an implicit cast to enforce a common type. Vectors can optionally be named, which results in a named matrix. Otherwise, default names are generated.

row_names
A character vector of row names for the resulting mtrx.

Details

mtrx objects are never reduced to vectors when subsetting using [ (i.e. dimensions are never dropped).

Examples

mtrx(1:10)

#> <vctrs_mtrx<integer>[,1][10]>
#>       [,1]
#>  [1,]  1  
#>  [2,]  2  
#>  [3,]  3  
#>  [4,]  4  
#>  [5,]  5  
#>  [6,]  6  
#>  [7,]  7  
#>  [8,]  8  
#>  [9,]  9  
#> [10,] 10  

a_mtrx <- mtrx(a = 1:5, b = 6:10)

a_mtrx

#> <vctrs_mtrx<integer>[,2][10]>
#>      a  b 
#> [1,]  1  6
#> [2,]  2  7
#> [3,]  3  8
#> [4,]  4  9
#> [5,]  5 10

# first column
a_mtrx[,1]

#> <vctrs_mtrx<integer>[,1][5]>
#>      a
#> [1,] 1
#> [2,] 2
#> [3,] 3
#> [4,] 4
#> [5,] 5

a_mtrx[,"a"]

#> <vctrs_mtrx<integer>[,1][5]>
#>      a
#> [1,] 1
#> [2,] 2
#> [3,] 3
#> [4,] 4
#> [5,] 5

# first row
a_mtrx[1,]

#> <vctrs_mtrx<integer>[,2][2]>
#>      a b
#> [1,] 1 6

# single element
a_mtrx[1, 1]

#> <vctrs_mtrx<integer>[,1][1]>
#>      a
#> [1,] 1

# multiple columns
a_mtrx[, c("a", "b")]

#> <vctrs_mtrx<integer>[,2][10]>
#>      a  b 
#> [1,]  1  6
#> [2,]  2  7
#> [3,]  3  8
#> [4,]  4  9
#> [5,]  5 10