Reducers

Apply a generic reducing function to the axes of .x. Reducing functions return a single value by iteratively applying a binary function. For example, rray_sum() is a reducing function that returns the sum along a specified axis.

rray_reduce_dbl(.x, .f, ..., axes = 1)

rray_reduce_int(.x, .f, ..., axes = 1)

rray_reduce_lgl(.x, .f, ..., axes = 1)

Arguments

.x
An array-like object.

.f
A 2-argument function. The function will be passed the accumulated value as the first argument and the "next" value as the second argument. It will be applied in this manner along each of the axes.

...
Additional arguments passed to .f.

axes
The axes to reduce over.

Examples

x <- rray(1:24, c(12, 2))

# Equivalent
rray_reduce_int(x, sum)

#> <vctrs_rray<integer>[,2][2]>
#>      [,1] [,2]
#> [1,]   78  222

rray_reduce_int(x, ~.x + .y)

#> <vctrs_rray<integer>[,2][2]>
#>      [,1] [,2]
#> [1,]   78  222

rray_sum(x)

#> <vctrs_rray<double>[,1][1]>
#>      [,1]
#> [1,]  300

# Same as above, but reduce the columns
rray_reduce_int(x, sum, axes = 2)

#> <vctrs_rray<integer>[,1][12]>
#>       [,1]
#>  [1,]   14
#>  [2,]   16
#>  [3,]   18
#>  [4,]   20
#>  [5,]   22
#>  [6,]   24
#>  [7,]   26
#>  [8,]   28
#>  [9,]   30
#> [10,]   32
#> [11,]   34
#> [12,]   36

rray_reduce_int(x, ~.x + .y, axes = 2)

#> <vctrs_rray<integer>[,1][12]>
#>       [,1]
#>  [1,]   14
#>  [2,]   16
#>  [3,]   18
#>  [4,]   20
#>  [5,]   22
#>  [6,]   24
#>  [7,]   26
#>  [8,]   28
#>  [9,]   30
#> [10,]   32
#> [11,]   34
#> [12,]   36

rray_sum(x, axes = 2)

#> <vctrs_rray<double>[,1][12]>
#>       [,1]
#>  [1,]   14
#>  [2,]   16
#>  [3,]   18
#>  [4,]   20
#>  [5,]   22
#>  [6,]   24
#>  [7,]   26
#>  [8,]   28
#>  [9,]   30
#> [10,]   32
#> [11,]   34
#> [12,]   36

# Running max of each column
x_shuffled <- x[c(1:5, 10:6, 12, 11),]
rray_reduce_int(x_shuffled, max)

#> <vctrs_rray<integer>[,2][2]>
#>      [,1] [,2]
#> [1,]   12   24

# Passing named arguments through
y <- x
y[1,1] <- NA
rray_reduce_int(y, sum, na.rm = TRUE)

#> <vctrs_rray<integer>[,2][2]>
#>      [,1] [,2]
#> [1,]   77  222

rray_reduce_int(y, ~sum(.x, .y, na.rm = TRUE))

#> <vctrs_rray<integer>[,2][2]>
#>      [,1] [,2]
#> [1,]   77  222