Find and count unique values in an array

rray_unique(): the unique values. Equivalent to unique() with the MARGIN argument.

rray_unique_loc(): the locations of the unique values.

rray_unique_count(): the number of unique values.

rray_unique(x, axis = 1L)

rray_unique_loc(x, axis = 1L)

rray_unique_count(x, axis = 1L)

Arguments

x
A vector, matrix, array, or rray.

axis
A single integer. The axis to look for unique values over.

Value

rray_unique(): an array the same type as x containing only unique values.

rray_unique_loc(): an integer vector, giving locations of the unique values.

rray_unique_count(): an integer vector of length 1, giving the number of unique values.

Details

When duplicates are detected, the first one is used in the result.

See also

Examples

x <- rray(c(1, 1, 2, 2), c(2, 2))
x <- set_row_names(x, c("r1", "r2"))
x <- set_col_names(x, c("c1", "c2"))

# Unique rows. The first unique
# row is used
rray_unique(x, 1L)

#> <rray<dbl>[,2][2]>
#>    c1 c2
#> r1  1  2

# Create a 3d version of x
# where the columns are not unique
y <- rray_expand_dims(x, 1)

# All of the rows are unique...
rray_unique(y, 1L)

#> <rray<dbl>[,2,2][4]>
#> , , c1
#> 
#>      r1 r2
#> [1,]  1  1
#> 
#> , , c2
#> 
#>      r1 r2
#> [1,]  2  2
#> 

# ...but the columns are not
rray_unique(y, 2L)

#> <rray<dbl>[,1,2][2]>
#> , , c1
#> 
#>      r1
#> [1,]  1
#> 
#> , , c2
#> 
#>      r1
#> [1,]  2
#> 

# The 3rd dimension is unique
rray_unique(y, 3L)

#> <rray<dbl>[,2,2][4]>
#> , , c1
#> 
#>      r1 r2
#> [1,]  1  1
#> 
#> , , c2
#> 
#>      r1 r2
#> [1,]  2  2
#> 

# rray_unique_loc() returns an
# integer vector you can use
# to subset out the unique values along
# the axis you are interested in
y[, rray_unique_loc(y, 2L)]

#> <rray<dbl>[,1,2][2]>
#> , , c1
#> 
#>      r1
#> [1,]  1
#> 
#> , , c2
#> 
#>      r1
#> [1,]  2
#> 

# Only 1 unique column
rray_unique_count(y, 2L)

#> [1] 1

# 2 unique 3D slices
rray_unique_count(y, 3L)

#> [1] 2